+++ src/ImGui/ImGuiWidgets.h
@@ -38,6 +38,27 @@ namespace ImGui
 		return retVal;
 	}
 
+	template <typename T, typename U, typename F>
+	bool SliderAutoFill(
+		const char*      label,
+		T*               v,
+		F                getter,
+		U                v_min,
+		U                v_max,
+		const char*      format = nullptr,
+		ImGuiSliderFlags flags  = 0)
+	{
+		if (v)
+		{
+			if (U* val = getter(v))
+				return ImGui::SliderAutoFill(label, val, v_min, v_max, format, flags);
+		}
+
+		U dummy = v_min;
+		ImGui::SliderAutoFill(label, &dummy, v_min, v_max, format, flags);
+		return false;
+	}
+
 	void NavBar(
 		const char* str_id,
 		const char* nav_item_id[],