+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Volumetric Lighting Editor
 * Fixed the menu size bug?
 * Moved camera options to new tab
+* Back lighting now ignores land